com32.h: Fix MK_PTR() for 64-bit
In the future we need 32/64-bit EFI to simply not include this file at
all. For now, just fixup the warning about casting integers to
pointers of different size.

+++ com32/include/com32.h
@@ -176,7 +176,7 @@ static inline bool _OFFS_VALID(const volatile void *__p, size_t __s,
 
 static inline void *MK_PTR(uint16_t __seg, uint16_t __offs)
 {
-    return (void *)((__seg << 4) + __offs);
+    return (void *)(unsigned long)((__seg << 4) + __offs);
 }
 
 /* Some tools to handle 16:16 far pointers in memory */
